+++ src/renderer/index.html
@@ -18,9 +18,40 @@
     />
 
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;600&display=swap" rel="stylesheet" />
-  </head>
 
+    <style>
+      html, body {
+        height: 100%;
+      }
+
+      body {
+        background-color: transparent !important;
+      }
+
+      .window-arrow-wrapper {
+        display: flex;
+        justify-content: center;
+        height:10px;
+
+        .window-arrow {
+          width: 0;
+          height: 0;
+          border-left: 10px solid transparent;
+          border-right: 10px solid transparent;
+          border-bottom: 10px solid #2d333b;
+        }
+      }
+
+      #root {
+        height: calc(100% - 10px);
+      }
+    </style>
+  </head>
   <body>
+    <div class="window-arrow-wrapper">
+      <div class="window-arrow"></div>
+    </div>
+
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
